passthrough: widen 32-bit change times
Convert Linux stat change-time fields at the platform boundary before deriving a QID version. On 32-bit Linux those fields are int32, while the shared nanosecond helper accepts int64.

+++ server/passthrough/stat_linux.go
@@ -48,7 +48,7 @@ func statToQID(st *unix.Stat_t) proto.QID {
 	}
 	return proto.QID{
 		Type:    t,
-		Version: qidVersion(st.Ctim.Sec, st.Ctim.Nsec),
+		Version: qidVersion(int64(st.Ctim.Sec), int64(st.Ctim.Nsec)),
 		Path:    st.Ino,
 	}
 }
